Version 1.3.16
* Add RefLocal, RefRemote and RefContext Schema Resolution Strategies

* Add Remotes to Test Suite

* Documentation
