+++ docs/docs/schema/1_spec.html
@@ -151,8 +151,8 @@
 <td align="left">-</td>
 <td align="left">-</td>
 <td align="left">-</td>
-<td align="left">38/44</td>
-<td align="left">19/27</td>
+<td align="left">40/44</td>
+<td align="left">21/27</td>
 </tr>
 <tr>
 <td align="left">enum</td>
@@ -412,7 +412,7 @@
 <td align="left">75/78</td>
 <td align="left">79/81</td>
 <td align="left">77/79</td>
-<td align="left">77/79</td>
+<td align="left">78/79</td>
 </tr>
 <tr>
 <td align="left">required</td>
@@ -514,8 +514,8 @@
 <td align="left">-</td>
 <td align="left">-</td>
 <td align="left">-</td>
-<td align="left">1/2</td>
-<td align="left">1/3</td>
+<td align="left">✅</td>
+<td align="left">2/3</td>
 <td align="left">0/1</td>
 <td align="left">-</td>
 </tr>
@@ -624,8 +624,8 @@
 <td align="left">✅</td>
 <td align="left">-</td>
 <td align="left">-</td>
-<td align="left">-</td>
-<td align="left">-</td>
+<td align="left">✅</td>
+<td align="left">✅</td>
 <td align="left">✅</td>
 <td align="left">✅</td>
 </tr>
