+++ test/jsonschema/cases/draft2019-09/optional/_cross-draft.json
@@ -1,25 +1,5 @@
 [
   {
-    "description": "refs to future drafts are processed as future drafts",
-    "schema": {
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "type": "array",
-      "$ref": "http://localhost:1234/draft2020-12/prefixItems.json"
-    },
-    "tests": [
-      {
-        "description": "first item is a string is valid",
-        "data": [
-          "a string",
-          1,
-          2,
-          3
-        ],
-        "valid": true
-      }
-    ]
-  },
-  {
     "description": "refs to historic drafts are processed as historic drafts",
     "schema": {
       "type": "object",
