+++ test/jsonschema/cases/draft2019-09/optional/cross-draft.json
@@ -16,6 +16,16 @@
           3
         ],
         "valid": false
+      },
+      {
+        "description": "first item is a string is valid",
+        "data": [
+          "a string",
+          1,
+          2,
+          3
+        ],
+        "valid": true
       }
     ]
   }
