+++ test/jsonschema/cases/draft2019-09/refRemote.json
@@ -7,6 +7,11 @@
     },
     "tests": [
       {
+        "description": "remote ref valid",
+        "data": 1,
+        "valid": true
+      },
+      {
         "description": "remote ref invalid",
         "data": "a",
         "valid": false
@@ -21,6 +26,11 @@
     },
     "tests": [
       {
+        "description": "remote fragment valid",
+        "data": 1,
+        "valid": true
+      },
+      {
         "description": "remote fragment invalid",
         "data": "a",
         "valid": false
@@ -34,6 +44,11 @@
       "$ref": "http://localhost:1234/draft2019-09/locationIndependentIdentifier.json#foo"
     },
     "tests": [
+      {
+        "description": "remote anchor valid",
+        "data": 1,
+        "valid": true
+      },
       {
         "description": "remote anchor invalid",
         "data": "a",
@@ -163,13 +178,11 @@
     },
     "tests": [
       {
-        "description": "object is invalid",
+        "description": "null is valid",
         "data": {
-          "name": {
-            "name": null
-          }
+          "name": null
         },
-        "valid": false
+        "valid": true
       }
     ]
   },
@@ -187,6 +200,13 @@
           "bar": 1
         },
         "valid": false
+      },
+      {
+        "description": "valid",
+        "data": {
+          "bar": "a"
+        },
+        "valid": true
       }
     ]
   },
@@ -238,6 +258,11 @@
         "description": "number is invalid",
         "data": 1,
         "valid": false
+      },
+      {
+        "description": "string is valid",
+        "data": "foo",
+        "valid": true
       }
     ]
   },
@@ -252,6 +277,11 @@
         "description": "number is invalid",
         "data": 1,
         "valid": false
+      },
+      {
+        "description": "string is valid",
+        "data": "foo",
+        "valid": true
       }
     ]
   },
