+++ test/jsonschema/cases/draft2020-12/_dynamicRef.json
@@ -1,61 +1,5 @@
 [
   {
-    "description": "strict-tree schema, guards against misspelled properties",
-    "schema": {
-      "$schema": "https://json-schema.org/draft/2020-12/schema",
-      "$id": "http://localhost:1234/draft2020-12/strict-tree.json",
-      "$dynamicAnchor": "node",
-      "$ref": "tree.json",
-      "unevaluatedProperties": false
-    },
-    "tests": [
-      {
-        "description": "instance with correct field",
-        "data": {
-          "children": [
-            {
-              "data": 1
-            }
-          ]
-        },
-        "valid": true
-      }
-    ]
-  },
-  {
-    "description": "tests for implementation dynamic anchor and reference link",
-    "schema": {
-      "$schema": "https://json-schema.org/draft/2020-12/schema",
-      "$id": "http://localhost:1234/draft2020-12/strict-extendible.json",
-      "$ref": "extendible-dynamic-ref.json",
-      "$defs": {
-        "elements": {
-          "$dynamicAnchor": "elements",
-          "properties": {
-            "a": true
-          },
-          "required": [
-            "a"
-          ],
-          "additionalProperties": false
-        }
-      }
-    },
-    "tests": [
-      {
-        "description": "correct extended schema",
-        "data": {
-          "elements": [
-            {
-              "a": 1
-            }
-          ]
-        },
-        "valid": true
-      }
-    ]
-  },
-  {
     "description": "$ref and $dynamicAnchor are independent of order - $defs first",
     "schema": {
       "$schema": "https://json-schema.org/draft/2020-12/schema",
@@ -82,15 +26,15 @@
     },
     "tests": [
       {
-        "description": "correct extended schema",
+        "description": "incorrect extended schema",
         "data": {
           "elements": [
             {
-              "a": 1
+              "b": 1
             }
           ]
         },
-        "valid": true
+        "valid": false
       }
     ]
   },
@@ -121,15 +65,15 @@
     },
     "tests": [
       {
-        "description": "correct extended schema",
+        "description": "incorrect extended schema",
         "data": {
           "elements": [
             {
-              "a": 1
+              "b": 1
             }
           ]
         },
-        "valid": true
+        "valid": false
       }
     ]
   },
