+++ test/jsonschema/cases/draft3/refRemote.json
@@ -6,6 +6,11 @@
     },
     "tests": [
       {
+        "description": "remote ref valid",
+        "data": 1,
+        "valid": true
+      },
+      {
         "description": "remote ref invalid",
         "data": "a",
         "valid": false
@@ -18,6 +23,11 @@
       "$ref": "http://localhost:1234/draft3/subSchemas.json#/definitions/integer"
     },
     "tests": [
+      {
+        "description": "remote fragment valid",
+        "data": 1,
+        "valid": true
+      },
       {
         "description": "remote fragment invalid",
         "data": "a",
