+++ test/jsonschema/cases/draft4/_refRemote.json
@@ -1,31 +1,5 @@
 [
   {
-    "description": "remote ref",
-    "schema": {
-      "$ref": "http://localhost:1234/integer.json"
-    },
-    "tests": [
-      {
-        "description": "remote ref valid",
-        "data": 1,
-        "valid": true
-      }
-    ]
-  },
-  {
-    "description": "fragment within remote ref",
-    "schema": {
-      "$ref": "http://localhost:1234/draft4/subSchemas.json#/definitions/integer"
-    },
-    "tests": [
-      {
-        "description": "remote fragment valid",
-        "data": 1,
-        "valid": true
-      }
-    ]
-  },
-  {
     "description": "ref within remote ref",
     "schema": {
       "$ref": "http://localhost:1234/draft4/subSchemas.json#/definitions/refToInteger"
