--- test/jsonschema/cases/draft7/optional/_cross-draft.json
@@ -1,28 +1,0 @@
-[
-  {
-    "description": "refs to future drafts are processed as future drafts",
-    "schema": {
-      "type": "object",
-      "allOf": [
-        {
-          "properties": {
-            "foo": true
-          }
-        },
-        {
-          "$ref": "http://localhost:1234/draft2019-09/dependentRequired.json"
-        }
-      ]
-    },
-    "tests": [
-      {
-        "description": "present bar is valid",
-        "data": {
-          "foo": "any value",
-          "bar": "also any value"
-        },
-        "valid": true
-      }
-    ]
-  }
-]
