+++ test/jsonschema/cases/draft7/optional/cross-draft.json
@@ -22,6 +22,14 @@
           "foo": "any value"
         },
         "valid": false
+      },
+      {
+        "description": "present bar is valid",
+        "data": {
+          "foo": "any value",
+          "bar": "also any value"
+        },
+        "valid": true
       }
     ]
   }
