+++ test/jsonschema/cases/v1/_dynamicRef.json
@@ -66,62 +66,6 @@
     ]
   },
   {
-    "description": "strict-tree schema, guards against misspelled properties",
-    "schema": {
-      "$schema": "https://json-schema.org/v1",
-      "$id": "http://localhost:1234/v1/strict-tree.json",
-      "$dynamicAnchor": "node",
-      "$ref": "tree.json",
-      "unevaluatedProperties": false
-    },
-    "tests": [
-      {
-        "description": "instance with correct field",
-        "data": {
-          "children": [
-            {
-              "data": 1
-            }
-          ]
-        },
-        "valid": true
-      }
-    ]
-  },
-  {
-    "description": "tests for implementation dynamic anchor and reference link",
-    "schema": {
-      "$schema": "https://json-schema.org/v1",
-      "$id": "http://localhost:1234/v1/strict-extendible.json",
-      "$ref": "extendible-dynamic-ref.json",
-      "$defs": {
-        "elements": {
-          "$dynamicAnchor": "elements",
-          "properties": {
-            "a": true
-          },
-          "required": [
-            "a"
-          ],
-          "additionalProperties": false
-        }
-      }
-    },
-    "tests": [
-      {
-        "description": "correct extended schema",
-        "data": {
-          "elements": [
-            {
-              "a": 1
-            }
-          ]
-        },
-        "valid": true
-      }
-    ]
-  },
-  {
     "description": "$ref and $dynamicAnchor are independent of order - $defs first",
     "schema": {
       "$schema": "https://json-schema.org/v1",
@@ -148,15 +92,15 @@
     },
     "tests": [
       {
-        "description": "correct extended schema",
+        "description": "incorrect extended schema",
         "data": {
           "elements": [
             {
-              "a": 1
+              "b": 1
             }
           ]
         },
-        "valid": true
+        "valid": false
       }
     ]
   },
@@ -187,15 +131,15 @@
     },
     "tests": [
       {
-        "description": "correct extended schema",
+        "description": "incorrect extended schema",
         "data": {
           "elements": [
             {
-              "a": 1
+              "b": 1
             }
           ]
         },
-        "valid": true
+        "valid": false
       }
     ]
   },
