+++ test/jsonschema/cases/v1/_ref.json
@@ -1,21 +1,5 @@
 [
   {
-    "description": "remote ref, containing refs itself",
-    "schema": {
-      "$schema": "https://json-schema.org/v1",
-      "$ref": "http://localhost:1234/v1/ref-and-defs.json"
-    },
-    "tests": [
-      {
-        "description": "remote ref valid",
-        "data": {
-          "bar": ""
-        },
-        "valid": true
-      }
-    ]
-  },
-  {
     "description": "order of evaluation: $id and $ref on nested schema",
     "schema": {
       "$comment": "$id must be evaluated before $ref to get the proper $ref destination",
