+++ test/jsonschema/cases/v1/_refRemote.json
@@ -1,47 +1,5 @@
 [
   {
-    "description": "remote ref",
-    "schema": {
-      "$schema": "https://json-schema.org/v1",
-      "$ref": "http://localhost:1234/v1/integer.json"
-    },
-    "tests": [
-      {
-        "description": "remote ref valid",
-        "data": 1,
-        "valid": true
-      }
-    ]
-  },
-  {
-    "description": "fragment within remote ref",
-    "schema": {
-      "$schema": "https://json-schema.org/v1",
-      "$ref": "http://localhost:1234/v1/subSchemas.json#/$defs/integer"
-    },
-    "tests": [
-      {
-        "description": "remote fragment valid",
-        "data": 1,
-        "valid": true
-      }
-    ]
-  },
-  {
-    "description": "anchor within remote ref",
-    "schema": {
-      "$schema": "https://json-schema.org/v1",
-      "$ref": "http://localhost:1234/v1/locationIndependentIdentifier.json#foo"
-    },
-    "tests": [
-      {
-        "description": "remote anchor valid",
-        "data": 1,
-        "valid": true
-      }
-    ]
-  },
-  {
     "description": "ref within remote ref",
     "schema": {
       "$schema": "https://json-schema.org/v1",
@@ -170,28 +128,13 @@
         "valid": true
       },
       {
-        "description": "null is valid",
+        "description": "object is invalid",
         "data": {
-          "name": null
+          "name": {
+            "name": null
+          }
         },
-        "valid": true
-      }
-    ]
-  },
-  {
-    "description": "remote ref with ref to defs",
-    "schema": {
-      "$schema": "https://json-schema.org/v1",
-      "$id": "http://localhost:1234/v1/schema-remote-ref-ref-defs1.json",
-      "$ref": "ref-and-defs.json"
-    },
-    "tests": [
-      {
-        "description": "valid",
-        "data": {
-          "bar": "a"
-        },
-        "valid": true
+        "valid": false
       }
     ]
   },
@@ -228,34 +171,6 @@
             "foo": "a"
           }
         },
-        "valid": true
-      }
-    ]
-  },
-  {
-    "description": "remote HTTP ref with different $id",
-    "schema": {
-      "$schema": "https://json-schema.org/v1",
-      "$ref": "http://localhost:1234/v1/different-id-ref-string.json"
-    },
-    "tests": [
-      {
-        "description": "string is valid",
-        "data": "foo",
-        "valid": true
-      }
-    ]
-  },
-  {
-    "description": "remote HTTP ref with different URN $id",
-    "schema": {
-      "$schema": "https://json-schema.org/v1",
-      "$ref": "http://localhost:1234/v1/urn-ref-string.json"
-    },
-    "tests": [
-      {
-        "description": "string is valid",
-        "data": "foo",
         "valid": true
       }
     ]
