+++ test/jsonschema/cases/v1/dynamicRef.json
@@ -218,6 +218,17 @@
           ]
         },
         "valid": false
+      },
+      {
+        "description": "instance with correct field",
+        "data": {
+          "children": [
+            {
+              "data": 1
+            }
+          ]
+        },
+        "valid": true
       }
     ]
   },
@@ -258,6 +269,17 @@
           ]
         },
         "valid": false
+      },
+      {
+        "description": "correct extended schema",
+        "data": {
+          "elements": [
+            {
+              "a": 1
+            }
+          ]
+        },
+        "valid": true
       }
     ]
   },
@@ -295,15 +317,15 @@
         "valid": false
       },
       {
-        "description": "incorrect extended schema",
+        "description": "correct extended schema",
         "data": {
           "elements": [
             {
-              "b": 1
+              "a": 1
             }
           ]
         },
-        "valid": false
+        "valid": true
       }
     ]
   },
@@ -341,15 +363,15 @@
         "valid": false
       },
       {
-        "description": "incorrect extended schema",
+        "description": "correct extended schema",
         "data": {
           "elements": [
             {
-              "b": 1
+              "a": 1
             }
           ]
         },
-        "valid": false
+        "valid": true
       }
     ]
   },
